python-interface-for-plc-communication/requirements.txt
# PLC Dashboard System - Requirements

# Core web frameworks
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP requests
requests>=2.31.0

# PLC communication
rk-mcprotocol==0.0.2

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Additional utilities
typing-extensions>=4.8.0
